# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP API for landlord application",
        "title": "Landlord API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/tenants": {
            "get": {
                "description": "Returns a paginated list of tenants",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenants"
                ],
                "summary": "List all tenants",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Maximum number of results (default 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results to skip (default 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include archived tenants in results",
                        "name": "include_deleted",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tenants",
                        "schema": {
                            "$ref": "#/definitions/models.ListTenantsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid pagination parameters",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new tenant with provided configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenants"
                ],
                "summary": "Create a new tenant",
                "parameters": [
                    {
                        "description": "Tenant creation request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateTenantRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tenant created successfully",
                        "schema": {
                            "$ref": "#/definitions/models.TenantResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or validation error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Tenant name already exists",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/tenants/{id}/archive": {
            "post": {
                "description": "Archives a tenant by removing compute resources and retaining the record",
                "tags": [
                    "tenants"
                ],
                "summary": "Archive a tenant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant identifier (UUID or name)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tenant already archived",
                        "schema": {
                            "$ref": "#/definitions/models.TenantResponse"
                        }
                    },
                    "202": {
                        "description": "Tenant archival initiated",
                        "schema": {
                            "$ref": "#/definitions/models.TenantResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid tenant identifier format",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Tenant not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Invalid state transition",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.CreateTenantRequest": {
            "type": "object",
            "required": [
                "image",
                "name"
            ],
            "properties": {
                "annotations": {
                    "description": "Annotations are key-value pairs for metadata",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "compute_config": {
                    "description": "ComputeConfig is provider-specific configuration (Docker, ECS, K8s, etc.)\nValidated by the active compute provider",
                    "type": "object",
                    "additionalProperties": true
                },
                "image": {
                    "description": "Image is the container image the tenant should run",
                    "type": "string"
                },
                "labels": {
                    "description": "Labels are key-value pairs for organizing tenants",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "Name is the unique human-friendly name for the tenant",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "Details provides additional context about the error",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "error": {
                    "description": "Error is the error message",
                    "type": "string"
                },
                "request_id": {
                    "description": "RequestID is the correlation ID from the request context",
                    "type": "string"
                }
            }
        },
        "models.ListTenantsResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "description": "Number of items per page",
                    "type": "integer"
                },
                "offset": {
                    "description": "Starting position",
                    "type": "integer"
                },
                "tenants": {
                    "description": "Tenants is the array of tenant resources",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TenantResponse"
                    }
                },
                "total": {
                    "description": "Pagination metadata",
                    "type": "integer"
                }
            }
        },
...
    }
}